+++ app.js
@@ -5,7 +5,8 @@ var cookieParser = require('cookie-parser');
 var logger = require('morgan');
 
 var indexRouter = require('./routes/index');
-var usersRouter = require('./routes/users');
+var testRouter = require('./routes/test');
+var test2Router = require('./routes/test2');
 
 var app = express();
 
@@ -20,7 +21,8 @@ app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', indexRouter);
-app.use('/users', usersRouter);
+app.use('/test', testRouter);
+app.use('/test2', test2Router);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
