+++ views/index.pug
@@ -2,4 +2,5 @@ extends layout
 
 block content
   h1= title
+  <h1>Hello, world!</h1>
   p Welcome to #{title}
